refactor: Separate ReplyHandler from CallWebSocketHandler
