+++ src/main/java/com/example/pitching/call/service/ActiveUserManager.java
@@ -1,4 +1,4 @@
-package com.example.pitching.call.handler;
+package com.example.pitching.call.service;
 
 import com.example.pitching.call.exception.DuplicateOperationException;
 import com.example.pitching.call.exception.ErrorCode;
